third_party/WebKit/LayoutTests/crypto/subtle/aes-cbc/encrypt-decrypt-expected.txt

Tests encrypt/decrypt for AES-CBC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-CBC"
PASS tmpKey.algorithm.length is 128
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [7649abac8119b246cee98e9b12e9197d5086cb9b507219ee95db113a917678b273bed6b8e3c1743b7116e69e222295163ff1caa1681fac09120eca307586e1a78cb82807230e1321d3fae00d18cc2012] and was
PASS: Decryption should be [6bc1bee22e409f96e93d7e117393172aae2d8a571e03ac9c9eb76fac45af8e5130c81c46a35ce411e5fbc1191a0a52eff69f2445df4f9b17ad2b417be66c3710] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-CBC"
PASS tmpKey.algorithm.length is 256
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [f58c4c04d6e5f1ba779eabfb5f7bfbd69cfc4e967edb808d679f777bc6702c7d39f23369a9d9bacfa530e26304231461c9aaf02a6a54e9e242ccbf48c59daca6] and was
PASS: Decryption should be [6bc1bee22e409f96e93d7e117393172aae2d8a571e03ac9c9eb76fac45af8e5130c81c46a35ce411e5fbc1191a0a52eff69f2445df4f9b17ad2b417be6] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "AES-CBC"
PASS tmpKey.algorithm.length is 128
PASS tmpKey.usages.join(',') is "encrypt,decrypt"
PASS: Encryption should be [8518b8878d34e7185e300d0fcc426396] and was
PASS: Decryption should be [] and was
PASS successfullyParsed is true

TEST COMPLETE
